Refactoring code of Armia.py and replacing armyID with reference to actual army.
